# apps/backend/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Helpmaton API",
    "version": "0.0.2",
    "description": "RESTful API for Helpmaton - AI agent workspace management and operations"
  },
  "servers": [
    {
      "url": "https://app.helpmaton.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:5173",
      "description": "Local Development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token authentication using JWT access tokens. Use the format: Authorization: Bearer <access-token>. Access tokens are obtained via /api/user/generate-tokens or /api/user/refresh-token."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "statusCode": 401,
              "error": "Unauthorized",
              "message": "Authentication required"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "statusCode": 403,
              "error": "Forbidden",
              "message": "Insufficient permissions"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "statusCode": 400,
              "error": "Bad Request",
              "message": "Invalid request parameters"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "statusCode": 500,
              "error": "Internal Server Error",
              "message": "An unexpected error occurred"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "statusCode",
          "error"
        ],
        "properties": {
          "statusCode": {
            "type": "integer",
            "description": "HTTP status code"
          },
          "error": {
            "type": "string",
            "description": "Error type"
          },
          "message": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
      "Workspace": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Workspace ID"
          },
          "name": {
            "type": "string",
            "description": "Workspace name"
          },
          "description": {
            "type": "string",
            "description": "Workspace description",
            "nullable": true
          },
          "permissionLevel": {
            "type": "integer",
            "description": "User's permission level in this workspace",
            "nullable": true
          },
          "creditBalance": {
            "type": "number",
            "description": "Current credit balance"
          },
          "currency": {
            "type": "string",
            "enum": [
              "usd",
              "eur",
              "gbp"
            ],
            "description": "Currency code"
          },
          "spendingLimits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpendingLimit"
            },
            "description": "Spending limits configuration"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          }
        }
      },
      "WorkspacesResponse": {
        "type": "object",
        "properties": {
          "workspaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Workspace"
            }
          }
        }
      },
      "WorkspaceResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/Workspace"
          }
        ]
      },
      "CreateWorkspaceRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Workspace name"
          },
          "description": {
            "type": "string",
            "description": "Workspace description"
          },
          "currency": {
            "type": "string",
            "enum": [
              "usd",
              "eur",
              "gbp"
            ],
            "description": "Currency code",
            "default": "usd"
          }
        }
      },
...
      "SpendingLimit": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "description": "Limit amount"
          },
          "currency": {
            "type": "string",
            "enum": [
              "usd",
              "eur",
              "gbp"
            ],
            "description": "Currency code"
          },
          "period": {
            "type": "string",
            "enum": [
              "daily",
              "weekly",
              "monthly"
            ],
            "description": "Time period"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/workspaces": {
      "post": {
        "summary": "Create a new workspace",
        "description": "Creates a new workspace for the authenticated user",
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workspace created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "get": {
        "summary": "List all workspaces",
        "description": "Returns all workspaces the authenticated user has access to",
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of workspaces",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspacesResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/workspaces/{workspaceId}/spending-limits": {
      "post": {
        "summary": "Create workspace spending limit",
        "description": "Creates a spending limit for a workspace to control costs over a specific time period (daily, weekly, or monthly). When the limit is reached, operations that would exceed it will be blocked. Multiple limits can be set for different time frames. Requires WRITE permission or higher.",
        "tags": [
          "Spending Limits"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "Workspace ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "timeFrame",
                  "amount"
                ],
                "properties": {
                  "timeFrame": {
                    "type": "string",
                    "enum": [
                      "daily",
                      "weekly",
                      "monthly"
                    ],
                    "description": "Time period for the spending limit"
                  },
                  "amount": {
                    "type": "number",
                    "minimum": 0,
                    "description": "Spending limit amount"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace with updated spending limits",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Workspace"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/workspaces/{workspaceId}/documents": {
      "post": {
        "summary": "Upload workspace documents",
        "description": "Uploads one or more documents (files or text) to a workspace. Supports up to 50 files. Allowed types are md, txt, and markdown.",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "Workspace ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Files to upload (max 10MB each, up to 50 files)"
                  },
                  "folderPath": {
                    "type": "string",
                    "description": "Optional folder path"
                  },
                  "textDocuments": {
                    "type": "string",
                    "description": "JSON array of text documents with name and content"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Documents uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "documents": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "filename": {
                            "type": "string"
                          },
                          "folderPath": {
                            "type": "string"
                          },
                          "contentType": {
                            "type": "string"
                          },
                          "size": {
                            "type": "integer"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "get": {
        "summary": "List workspace documents",
        "description": "Returns all documents in a workspace, optionally filtered by folder",
        "tags": [
          "Documents"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "description": "Workspace ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "folder",
            "in": "query",
            "description": "Optional folder path to filter documents",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "documents": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "filename": {
                            "type": "string"
                          },
                          "folderPath": {
                            "type": "string"
                          },
                          "contentType": {
                            "type": "string"
                          },
                          "size": {
                            "type": "integer"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  }
}